Get pty working and create POC
The POC works only right now with echoing enabled for the expect-
like DSL. This is undesirable but the best I can seem to do with
the current limitations in the Crystal language. The next thing
to try will be changing things so that we imitate how the C code
would look internally by calling LibC.fork directly.