Update logic to properly create process pty
This should be opened with the LibC open command so that it's in
tty mode and can be turned into raw tty mode. Unfortunately that
is not possible yet because there is a bug in the current Crystal
version with turning file descriptors into raw mode without a block.
That should be fixed in Crystal 1.13.0.

See https://github.com/crystal-lang/crystal/pull/14529